Unifico rutas de usuarios, controlador de usuarios
