+++ app.js
@@ -7,10 +7,9 @@ let logger = require('morgan');
 const expressSession = require("express-session");
 
 let indexRouter = require('./src/routes/index');
-let cartRouter = require('./src/routes/cart');
-let loginRouter = require('./src/routes/login');
-let registerRouter = require('./src/routes/register');
-let productRouter = require('./src/routes/product');
+// let loginRouter = require('./src/routes/login');
+// let registerRouter = require('./src/routes/register');
+let productRouter = require('./src/routes/productRouter');
 let userRouter = require('./src/routes/userRouter');
 
 let userLoggedMiddleware = require('./src/middlewares/userLoggedMiddleware')
@@ -31,10 +30,9 @@ app.use(expressSession({secret: "secret",resave: false,saveUninitialized: false}
 app.use(userLoggedMiddleware);
 
 app.use('/', indexRouter);
-app.use('/cart', cartRouter);
-app.use('/login', loginRouter);
-app.use('/register', registerRouter);
-app.use('/product', productRouter);
+// app.use('/login', loginRouter);
+// app.use('/register', registerRouter);
+app.use('/products', productRouter);
 app.use('/users', userRouter);
 
 // catch 404 and forward to error handler
