+++ src/database/users.json
@@ -108,5 +108,25 @@
   "adress": "calle de colombia",
   "image": "/images/users/1671419690187.ico",
   "id": 11
+ },
+ {
+  "names": "prueba",
+  "surnames": "qwerty",
+  "password": "$2a$10$AUMan7Cfvtxn2YrOFrO5vOhY8Np85zj4Xq99Pc/8r048GGQdn0c26",
+  "phone": "54984546845",
+  "email": "prueba@qwerty.com",
+  "adress": "uhssdhAKJHJK",
+  "image": "/images/users/1671890562168.ico",
+  "id": 12
+ },
+ {
+  "names": "prueba123",
+  "surnames": "123",
+  "password": "$2a$10$6mZR9QM25.sj4f6utnelzu4NrQ1SU91C4ccsUmO8Sma.FTAVjSrH.",
+  "phone": "123456789",
+  "email": "prueba@qwerty2.com",
+  "adress": "fyftsertaedfds",
+  "image": "/images/users/1671893186231.ico",
+  "id": 13
  }
 ]
