+++ src/routes/index.js
@@ -1,12 +1,13 @@
 const express = require('express');
 const router = express.Router();
 const mainController =require('../controllers/mainController');
+const productController =require('../controllers/productController');
 
-router.get('/', mainController.index);
-router.get('/best-selling', mainController.best);
-router.get('/offers', mainController.offer);
-router.get('/recommended', mainController.recommended);
-router.get('/products', mainController.allProducts);
+router.get('/', productController.index);
+// router.get('/best-selling', mainController.best);
+// router.get('/offers', mainController.offer);
+// router.get('/recommended', mainController.recommended);
+// router.get('/products', mainController.allProducts);
 
 
 module.exports=router;
