+++ src/views/partials/header.ejs
@@ -6,19 +6,19 @@
       </div>
       <ul class="header-user-list">
         <% if (!locals.isLogged){%>
-          <li><a class="header-link" href="users/login">Login</a></li>
+          <li><a class="header-link" href="/users/login">Login</a></li>
         <%}%>
         <% if (locals.isLogged){%>
           <li><a class="header-link" href="/users/logout">Logout</a></li>
         <%}%>
         <% if (!locals.isLogged){%>
-          <li><a class="header-link" href="users/register">Register</a></li>
+          <li><a class="header-link" href="/users/register">Register</a></li>
         <%}%>
         <% if (locals.isLogged){%>
           <li class="header-name"><a href="/users/detail/<%=locals.userLogged.id%>"><%=locals.userLogged.names%></a></li>
         <%}%>
         <li>
-          <a class="header-icon" href="/cart"
+          <a class="header-icon" href="/products/cart/all"
             ><em class="fa-solid fa-cart-shopping"></em
           ></a>
         </li>
@@ -31,10 +31,10 @@
       <nav>
         <ul class="header-nav-bar">
           <li><a class="header-link" href="/">Tienda</a></li>
-          <li><a class="header-link" href="/offers">Ofertas</a></li>
-          <li><a class="header-link" href="/recommended">Recomendados</a></li>
-          <li><a class="header-link" href="/products">Categorias</a></li>
-          <li><a class="header-link" href="/best-selling">Más vendidos</a></li>
+          <li><a class="header-link" href="/products/offers">Ofertas</a></li>
+          <li><a class="header-link" href="/products/recommended">Recomendados</a></li>
+          <li><a class="header-link" href="/products/all">Categorias</a></li>
+          <li><a class="header-link" href="/products//best-selling">Más vendidos</a></li>
           <li>
             <div class="search">
               <form action="#">
