+++ src/views/products/editProduct.ejs
@@ -11,7 +11,7 @@
            
               <img src=" <%= product.imageUri %> "  class="edit-product-image" alt="imagen de juego" >
            
-            <form action="/product/<%=product.id%>/edit?_method=PUT" class="edit-product-form" method="post"
+            <form action="/products/<%=product.id%>/edit?_method=PUT" class="edit-product-form" method="post"
               enctype="multipart/form-data">
 
               <div class="edit-product-div">
