+++ src/views/products/main.ejs
@@ -33,7 +33,7 @@
                                         <h4><%=products[i].name%></h4>
                                         <p>Categoría: <%=products[i].category%></p>
                                         <div class="link-buy">
-                                            <a href="/product/<%=products[i].id%>"><%=products[i].value %> - COMPRAR</a>
+                                            <a href="/products/<%=products[i].id%>"><%=products[i].value %> - COMPRAR</a>
                                         </div>
                                     </div>
                                 </div>
@@ -65,7 +65,7 @@
                                         <h4><%=products[i].name%></h4>
                                         <p>Categoría: <%=products[i].category%></p>
                                         <div class="link-buy">
-                                            <a href="/product/<%=products[i].id%>"><%=products[i].value %> - COMPRAR</a>
+                                            <a href="/products/<%=products[i].id%>"><%=products[i].value %> - COMPRAR</a>
                                         </div>
                                     </div>
                                 </div>
@@ -97,7 +97,7 @@
                                         <h4><%=products[i].name%></h4>
                                         <p>Categoría: <%=products[i].category%></p>
                                         <div class="link-buy">
-                                            <a href="/product/<%=products[i].id%>"><%=products[i].value %> - COMPRAR</a>
+                                            <a href="/products/<%=products[i].id%>"><%=products[i].value %> - COMPRAR</a>
                                         </div>
                                     </div>
                                 </div>
