+++ src/views/products/newProduct.ejs
@@ -7,7 +7,7 @@
       <div class="new-product-box">
         <h3>Agregar nuevo producto</h3>
         <div class="new-product-form-box">
-        <form action="/product/new" class="new-product-form" method="post" enctype="multipart/form-data">
+        <form action="/products/new" class="new-product-form" method="post" enctype="multipart/form-data">
           <div class="new-product-div">
             <label for="name" class="new-product-label">Nombre: </label>
             <input type="text" name="name" id="name"
