+++ src/views/products/recommended.ejs
@@ -25,7 +25,7 @@
                                 <h4><%=products[i].name%></h4>
                                 <p>Categoría: <%=products[i].category%></p>
                                 <div class="link-buy">
-                                    <a href="/product/<%=products[i].id%>"><%=products[i].value %> - COMPRAR</a>
+                                    <a href="/products/<%=products[i].id%>"><%=products[i].value %> - COMPRAR</a>
                                 </div>
                             </div>
                         </div>
