+++ src/views/register/register.ejs
@@ -8,7 +8,7 @@
     <div class="register-box">
       <h3 class="title">Registro</h3>
       <hr />
-      <form class="register-form" action="/register" method="POST" enctype="multipart/form-data">
+      <form class="register-form" action="/users/register" method="POST" enctype="multipart/form-data">
         <div class="input-box">
           <div class="register-block-inline">
             <div class="register-division">
