+++ src/views/users/login.ejs
@@ -7,7 +7,7 @@
     <main class="main-login">
         <div class="login-box">
             <h3>Login</h3>
-            <form class="login-form" action="/login" method="post">
+            <form class="login-form" action="/users/login" method="post">
                 <label class='label-login' for="email">Email</label>
                 <input type="email" name="email" id="email" required placeholder="name@mail.com">
                 <% if (locals.errors && errors.email){%>
@@ -29,7 +29,7 @@
                 <button class="buy-button" type="submit">Ingresar</button>
                 <a href="#" class="forgot-pass">¿Olvidaste tu contraseña?</a>
             </form>
-            <a class="register" href="register">Registrarse</a>
+            <a class="register" href="/users/register">Registrarse</a>
         </div>
     </main>
     <%- include('../partials/footer')%>
